feat: list in style
