+++ studio/src/app/popovers/editor/actions/app-more-element-actions/app-more-element-actions.tsx
@@ -14,9 +14,6 @@ export class AppMoreElementActions {
   @Prop()
   copy: boolean = false;
 
-  @Prop()
-  list: boolean = false;
-
   private async closePopover(action: MoreAction) {
     await (this.el.closest('ion-popover') as HTMLIonPopoverElement).dismiss({
       action: action,
@@ -28,7 +25,6 @@ export class AppMoreElementActions {
       <div class="ion-padding">
         {this.renderNotes()}
         {this.renderCopy()}
-        {this.renderList()}
         {this.renderDelete()}
       </div>
     );
@@ -54,18 +50,6 @@ export class AppMoreElementActions {
     return (
       <a onClick={() => this.closePopover(MoreAction.COPY)} aria-label="Copy">
         <p>Copy</p>
-      </a>
-    );
-  }
-
-  private renderList() {
-    if (!this.list) {
-      return undefined;
-    }
-
-    return (
-      <a onClick={() => this.closePopover(MoreAction.LIST)} aria-label="List">
-        <p>List</p>
       </a>
     );
   }
